+++ agregarEntregaCredito.php
@@ -7,6 +7,9 @@ if (isset($_GET["id"])) {
     $result_facturacion = mysqli_query($conn, $query);
     $contador = 0; //controlo la descripcion del producto
     $total = 0;
+    $productosArray = array();
+    $totalArray = array();
+    $cantidadArray = array();
 
     while ($row = mysqli_fetch_array($result_facturacion)) {
         $contador++;
@@ -20,6 +23,9 @@ if (isset($_GET["id"])) {
         $precioVenta = $row['precioVenta'];
         $subTotal = $row['subtotal'];
         $total = $total + $subTotal;
+        array_push($productosArray, "$producto");
+        array_push($totalArray, "$subTotal");
+        array_push($cantidadArray, "$cantidad");
 
         if ($contador == 1) {
             $idDetalle = uniqid(); //AQUI SE GENERA UN ID UNICO PARA LA FACTURACION
@@ -31,18 +37,30 @@ if (isset($_GET["id"])) {
         $query = "INSERT INTO detalleentregacredito(vendedor,cliente, producto, cantidad, precioCompra,precioVenta,idDetalle,idVendedor,idCliente,subtotal) VALUES('$vendedor','$cliente','$producto','$cantidad','$precioCompra', '$precioVenta','$idDetalle','$idVendedor','$idCliente','$subTotal')";
         mysqli_query($conn, $query);
     }
-    $query = "INSERT INTO entregamuestracredito(vendedor,cliente, descripcion,idDetalle,idvendedor,idCliente,total,estado) VALUES('$vendedor','$cliente','$descripcion','$idDetalle','$idVendedor','$idCliente','$total',1)";
-    mysqli_query($conn, $query);
 
-    $queryDelete = "DELETE from nuevaentregacredito";
-    mysqli_query($conn, $queryDelete);
+    if ($contador != 0) {
+        $query = "INSERT INTO entregamuestracredito(vendedor,cliente, descripcion,idDetalle,idvendedor,idCliente,total,estado) VALUES('$vendedor','$cliente','$descripcion','$idDetalle','$idVendedor','$idCliente','$total',1)";
+        mysqli_query($conn, $query);
 
-    $queryHistorial = "INSERT INTO historial(descripcion) VALUES('$_SESSION[usuario] Registró Un Creditos De Productos Al Cliente $cliente')";
-    mysqli_query($conn, $queryHistorial);
+        $queryDelete = "DELETE from nuevaentregacredito";
+        mysqli_query($conn, $queryDelete);
 
-    $_SESSION['EntregaAgregadaCredito'] = 1;
+        $queryHistorial = "INSERT INTO historial(descripcion) VALUES('$_SESSION[usuario] Registró Un Creditos De Productos Al Cliente $cliente')";
+        mysqli_query($conn, $queryHistorial);
 
-    echo $idDetalle;
+        $queryCliente = "SELECT * from clientes WHERE id= $idCliente";
+        $resultCliente = mysqli_query($conn, $queryCliente);
+        $rowCliente = mysqli_fetch_array($resultCliente);
+
+        $_SESSION['EntregaAgregadaCredito'] = 1;
+        $_SESSION['nombreCliente'] = $rowCliente['nombre'];
+        $_SESSION['telefonoCliente'] = $rowCliente['telefono'];
+        $_SESSION['productos'] = $productosArray;
+        $_SESSION['total'] = $totalArray;
+        $_SESSION['cantidades'] = $cantidadArray;
+    }
+
+
 ?>
     <script>
         window.location = "creditos.php"
